feat(images): address PR review feedback - shared types, robust detection, and test IDs
